Feat: New RoomService. Change Stay to have start and end dates instead of the latter being length
